requirements.txt
accelerate==1.0.1
albumentations==2.0.1
bitsandbytes==0.44.1
clip==1.0
diffusers==0.29.2
einops==0.8.0
fire==0.7.0
gradio==5.13.1
huggingface_hub==0.26.2
ipython==8.12.3
kornia==0.8.0
matplotlib==3.10.0
natsort==8.4.0
nitro==0.0.0
numpy==2.2.2
omegaconf==2.3.0
onnxruntime==1.20.1
opencv_python==4.10.0.84
packaging==24.2
pandas==2.2.3
Pillow==11.1.0
psutil==6.1.0
pytorch_lightning==2.5.0.post0
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
safetensors==0.5.2
scipy==1.15.1
simple_parsing==0.1.7
six==1.16.0
tensorflow==2.18.0
tensorflow_gan==2.1.0
tensorflow_hub==0.16.1
terminaltables==3.1.10
timm==1.0.14
torch==2.5.1
torchmetrics==1.6.0
torchvision==0.20.1
tqdm==4.66.6
transformers==4.46.1
triton==3.1.0
wandb==0.19.4
webdataset==0.2.100
xformers==0.0.29.post1
